Refactored the feature into a new component. Removed unused props from password fields.
