+++ frontend/src/components/board/AccessBoardForm.tsx
@@ -1,20 +1,14 @@
-import Visibility from "@mui/icons-material/Visibility"
-import VisibilityOff from "@mui/icons-material/VisibilityOff"
 import Button from "@mui/material/Button"
 import Divider from "@mui/material/Divider"
-import FormControl from "@mui/material/FormControl"
-import FormHelperText from "@mui/material/FormHelperText"
 import Grid from "@mui/material/Grid"
-import IconButton from "@mui/material/IconButton"
-import InputAdornment from "@mui/material/InputAdornment"
-import InputLabel from "@mui/material/InputLabel"
-import OutlinedInput from "@mui/material/OutlinedInput"
 import Typography from "@mui/material/Typography"
 import React from "react"
 import { useForm } from "react-hook-form"
 import { useNavigate } from "react-router-dom"
 
 import { useLoginMutation } from "@/state/apiSlice"
+
+import PasswordField from "./PasswordField"
 
 interface AccessBoardFormProps {
   id: string
@@ -25,11 +19,7 @@ interface FormData {
 }
 
 const AccessBoardForm: React.FC<AccessBoardFormProps> = ({ id }) => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors }
-  } = useForm<FormData>({
+  const { register, handleSubmit } = useForm<FormData>({
     defaultValues: {
       password: ""
     }
@@ -59,18 +49,6 @@ const AccessBoardForm: React.FC<AccessBoardFormProps> = ({ id }) => {
     onSubmit(data)
   }
 
-  const [showPassword, setShowPassword] = React.useState(false)
-
-  const handleClickShowPassword = () => setShowPassword((show) => !show)
-
-  const handleMouseDownPassword = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault()
-  }
-
-  const handleMouseUpPassword = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault()
-  }
-
   return (
     <form onSubmit={handleSubmit(handleFormSubmit)}>
       <Grid container spacing={1}>
@@ -81,28 +59,7 @@ const AccessBoardForm: React.FC<AccessBoardFormProps> = ({ id }) => {
           <Divider />
         </Grid>
         <Grid item xs={12}>
-          <FormControl>
-            <InputLabel htmlFor="outlined-adornment-password">Password</InputLabel>
-            <OutlinedInput
-              label="Password"
-              type={showPassword ? "Text" : "Password"}
-              error={Boolean(errors.password)}
-              {...register("password")}
-              endAdornment={
-                <InputAdornment position="end">
-                  <IconButton
-                    onClick={handleClickShowPassword}
-                    onMouseDown={handleMouseDownPassword}
-                    onMouseUp={handleMouseUpPassword}
-                    edge="end"
-                  >
-                    {showPassword ? <VisibilityOff /> : <Visibility />}
-                  </IconButton>
-                </InputAdornment>
-              }
-            />
-            <FormHelperText>{errors.password?.message}</FormHelperText>
-          </FormControl>
+          <PasswordField register={register("password")} />
         </Grid>
         <Grid item xs={12}>
           <Button type="submit" color="primary" variant="contained">
