+++ frontend/src/components/board/BoardCreationForm.tsx
@@ -1,19 +1,13 @@
-import Visibility from "@mui/icons-material/Visibility"
-import VisibilityOff from "@mui/icons-material/VisibilityOff"
 import Button from "@mui/material/Button"
 import Divider from "@mui/material/Divider"
-import FormControl from "@mui/material/FormControl"
 import Grid from "@mui/material/Grid"
-import IconButton from "@mui/material/IconButton"
-import InputAdornment from "@mui/material/InputAdornment"
-import InputLabel from "@mui/material/InputLabel"
-import OutlinedInput from "@mui/material/OutlinedInput"
 import TextField from "@mui/material/TextField"
 import Typography from "@mui/material/Typography"
-import React from "react"
 import { useForm } from "react-hook-form"
 
 import { NewBoardFormData } from "../../types"
+
+import PasswordField from "./PasswordField"
 
 interface AddBoardCreationFormProps {
   onSubmit: (_: NewBoardFormData) => void
@@ -31,18 +25,6 @@ const BoardCreationForm: React.FC<AddBoardCreationFormProps> = ({ onSubmit, onCa
       password: ""
     }
   })
-
-  const [showPassword, setShowPassword] = React.useState(false)
-
-  const handleClickShowPassword = () => setShowPassword((show) => !show)
-
-  const handleMouseDownPassword = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault()
-  }
-
-  const handleMouseUpPassword = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault()
-  }
 
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
@@ -83,27 +65,8 @@ const BoardCreationForm: React.FC<AddBoardCreationFormProps> = ({ onSubmit, onCa
             })}
           />
         </Grid>
-        <Grid item xs={9.75}>
-          <FormControl>
-            <InputLabel htmlFor="outlined-adornment-password">Password</InputLabel>
-            <OutlinedInput
-              type={showPassword ? "Text" : "Password"}
-              {...register("password")}
-              endAdornment={
-                <InputAdornment position="end">
-                  <IconButton
-                    onClick={handleClickShowPassword}
-                    onMouseDown={handleMouseDownPassword}
-                    onMouseUp={handleMouseUpPassword}
-                    edge="end"
-                  >
-                    {showPassword ? <VisibilityOff /> : <Visibility />}
-                  </IconButton>
-                </InputAdornment>
-              }
-              label="Password"
-            />
-          </FormControl>
+        <Grid item xs={12}>
+          <PasswordField register={register("password")} />
         </Grid>
         <Grid item xs={12}>
           <Button type="submit" color="primary" variant="contained">
